feat(Schedule): 일정 조회시 댓글 조회 기능 추가
데이터 접근 권한 확인하는 기능 추가

+++ src/main/java/com/example/scheduleprojectver2/service/ScheduleService.java
@@ -1,24 +1,33 @@
 package com.example.scheduleprojectver2.service;
 
+import com.example.scheduleprojectver2.dto.comments.ScheduleInCommentDto;
+import com.example.scheduleprojectver2.dto.schedules.SchedulePageReponseDto;
 import com.example.scheduleprojectver2.dto.schedules.ScheduleResponseDto;
 import com.example.scheduleprojectver2.dto.users.UserResponseDto;
 
 import com.example.scheduleprojectver2.entity.ScheduleEntity;
 import com.example.scheduleprojectver2.entity.UserEntity;
 import com.example.scheduleprojectver2.exception.ErrorDtoException;
+import com.example.scheduleprojectver2.exception.ForbiddenException;
+import com.example.scheduleprojectver2.repository.CommentRepository;
 import com.example.scheduleprojectver2.repository.ScheduleRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.List;
+
 @Service
 @RequiredArgsConstructor
 public class ScheduleService {
 
     private final ScheduleRepository scheduleRepository;
+
+    private final CommentRepository commentRepository;
 
     private final UserService userService;
 
@@ -40,7 +49,13 @@ public class ScheduleService {
 
         ScheduleEntity findSchedule = scheduleRepository.findByIdOrElseThrow(id);
 
-        return findSchedule.toDto();
+        ScheduleResponseDto responseDto = findSchedule.toDto();
+
+        List<ScheduleInCommentDto> comments = commentRepository.findCommentsByScheduleId(findSchedule.getId());
+
+        responseDto.setComments(comments);
+
+        return responseDto;
     }
 
 
@@ -48,20 +63,33 @@ public class ScheduleService {
         return scheduleRepository.findByIdOrElseThrow(id);
     }
 
-    public Page<ScheduleResponseDto> findAll(int page, int size) {
+    public Page<SchedulePageReponseDto> findAll(int page, int size) {
 
         PageRequest pageRequest = PageRequest.of(page, size, Sort.by(Sort.Direction.DESC, "updateDate"));
 
         return scheduleRepository.findAll(pageRequest)
-                .map(ScheduleEntity::toDto);
+                .map( item -> SchedulePageReponseDto.builder()
+                                .id(item.getId())
+                                .username(item.getUser().getUsername())
+                                .title(item.getTitle())
+                                .todo(item.getTodo())
+                                .commentCount(commentRepository.countBySchedule_Id(item.getId()))
+                                .build());
 
     }
 
     // 일괄 수정
     @Transactional
-    public ScheduleResponseDto update(Long id, String title, String todo) {
+    public ScheduleResponseDto update(Long id, Long userId, String title, String todo) {
+
+        UserEntity user = userService.findByIdToEntity(userId);
 
         ScheduleEntity findSchedule = scheduleRepository.findByIdOrElseThrow(id);
+
+        // 수정할 수 있는 권한이 없음
+        if( user != findSchedule.getUser() ) {
+            throw new ForbiddenException();
+        }
 
         findSchedule.updateTitle(title);
         findSchedule.updateTodo(todo);
@@ -71,12 +99,19 @@ public class ScheduleService {
 
     // 부분 수정
     @Transactional
-    public ScheduleResponseDto patch(Long id, String title, String todo) {
+    public ScheduleResponseDto patch(Long id, Long userId, String title, String todo) {
         if( (title.isEmpty() && todo.isEmpty()) || (!title.isEmpty() && !todo.isEmpty()) ) { // 모두 있거나, 모두 비어있거나
             throw new ErrorDtoException("잘못된 입력입니다.");
         }
 
+        UserEntity user = userService.findByIdToEntity(userId);
+
         ScheduleEntity findSchedule = scheduleRepository.findByIdOrElseThrow(id);
+
+        // 수정할 수 있는 권한이 없음
+        if( user != findSchedule.getUser() ) {
+            throw new ForbiddenException();
+        }
 
         if( !title.isEmpty() ) {
             findSchedule.updateTitle(title);
@@ -90,8 +125,16 @@ public class ScheduleService {
 
     // 일정 조회후 삭제
     @Transactional
-    public void delete(Long id) {
+    public void delete(Long id, Long userId) {
+
+        UserEntity user = userService.findByIdToEntity(userId);
+
         ScheduleEntity findSchedule = scheduleRepository.findByIdOrElseThrow(id);
+
+        // 수정할 수 있는 권한이 없음
+        if( user != findSchedule.getUser() ) {
+            throw new ForbiddenException();
+        }
 
         scheduleRepository.delete(findSchedule);
     }
